+++ Devantler.KindCLI.Tests.Integration/KindCliServiceTests/LoadImageArchiveAsyncTests.cs
@@ -2,15 +2,35 @@ namespace Devantler.KindCLI.Tests.Integration.KindCliServiceTests;
 
 public class LoadImageArchiveAsync : KindCliServiceTestsBase
 {
-    [Fact]
-    public async Task LoadImageArchiveAsync_ValidImageArchivePathAndValidClusterName_LoadsImageArchiveToNamedCluster()
+    public static IEnumerable<object[]> ValidCases
+    {
+        get
+        {
+            yield return new object[] { Guid.NewGuid().ToString(), "assets/nginx.tar" };
+        }
+    }
+
+    public static IEnumerable<object[]> InvalidCases
+    {
+        get
+        {
+            yield return new object[] { "@_~", "assets/nginx.tar", ErrorMessages.InvalidClusterName("@_~") };
+            yield return new object[] { "@_~", "assets/does-not-exist.tar", ErrorMessages.InvalidClusterName("@_~") };
+            yield return new object[] { "@_~", null!, ErrorMessages.InvalidClusterName("@_~") };
+            yield return new object[] { Guid.NewGuid().ToString(), null!, ErrorMessages.NullReference() };
+            yield return new object[] { Guid.NewGuid().ToString(), "assets/does-not-exist.tar", ErrorMessages.InvalidImageArchivePath("assets/does-not-exist.tar") };
+        }
+    }
+
+    [Theory]
+    [MemberData(nameof(ValidCases))]
+    public async Task LoadImageArchiveAsync_ValidParameters_LoadsImageArchiveToNamedCluster(string? clusterName, string imageArchivePath)
     {
         //Arrange
-        const string imageArchivePath = "assets/nginx.tar";
-        string clusterName = Guid.NewGuid().ToString();
+        if (clusterName != "@_~")
+            await _kindCliService.CreateClusterAsync(clusterName);
 
         //Act
-        await _kindCliService.CreateClusterAsync(clusterName);
         Func<Task> action = async () => await _kindCliService.LoadImageArchiveAsync(imageArchivePath, clusterName);
 
         //Assert
@@ -20,52 +40,22 @@ public class LoadImageArchiveAsync : KindCliServiceTestsBase
         _ = await _kindCliService.DeleteClusterAsync(clusterName);
     }
 
-    [Fact]
-    public void LoadImageArchiveAsync_ValidImageArchivePathAndInvalidClusterName_Throws()
+    [Theory]
+    [MemberData(nameof(InvalidCases))]
+    public async Task LoadImageArchiveAsync_InvalidParameters_ThrowsAsync(string? clusterName, string imageArchivePath, string expected)
     {
         //Arrange
-        const string imageArchivePath = "assets/nginx.tar";
-        const string clusterName = "@_~";
-        const string expected = $"The specified 'clusterName': '{clusterName}' is invalid. It must match '^[a-z0-9.-]+$'";
+        if (clusterName != "@_~")
+            await _kindCliService.CreateClusterAsync(clusterName);
 
         //Act
-        Func<Task> action = async () => await _kindCliService.LoadImageArchiveAsync(imageArchivePath, clusterName);
-
-        //Assert
-        _ = action.Should().ThrowAsync<ArgumentException>().WithMessage(expected);
-    }
-
-    [Fact]
-    public async Task LoadImageArchiveAsync_InvalidImageArchivePathAndValidClusterName_Throws()
-    {
-        //Arrange
-        const string imageArchivePath = "invalid-path/nginx.tar";
-        string clusterName = Guid.NewGuid().ToString();
-        const string expected = $"The specified 'imageArchivePath': '{imageArchivePath}' does not exist.";
-
-        //Act
-        await _kindCliService.CreateClusterAsync(clusterName);
         Func<Task> action = async () => await _kindCliService.LoadImageArchiveAsync(imageArchivePath, clusterName);
 
         //Assert
         _ = action.Should().ThrowAsync<ArgumentException>().WithMessage(expected);
 
         //Cleanup
-        _ = await _kindCliService.DeleteClusterAsync(clusterName);
-    }
-
-    [Fact]
-    public void LoadImageArchiveAsync_InvalidImageArchivePathAndInvalidClusterName_Throws()
-    {
-        //Arrange
-        const string imageArchivePath = "invalid-path/nginx.tar";
-        const string clusterName = "@_~";
-        const string expected = $"The specified 'clusterName': '{clusterName}' is invalid. It must match '^[a-z0-9.-]+$'";
-
-        //Act
-        Func<Task> action = async () => await _kindCliService.LoadImageArchiveAsync(imageArchivePath, clusterName);
-
-        //Assert
-        _ = action.Should().ThrowAsync<ArgumentException>().WithMessage(expected);
+        if (clusterName != "@_~")
+            _ = await _kindCliService.DeleteClusterAsync(clusterName);
     }
 }
